updating the entities id's to use long instead of int as some identities returned are int64

+++ Entities/Attendee.cs
@@ -9,9 +9,9 @@ namespace EventbriteNET.Entities
     }
     public class Attendee : EntityBase
     {
-        public int Id;
-        public int EventId;
-        public int? TicketId;
+        public long Id;
+        public long EventId;
+        public long? TicketId;
         public int? Quantity;
         public string Currency;
         public float? AmountPaid;
